Use JUnit's TemporaryFolder instead of deleting files in tearDown().
