+++ src/test/java/org/eclipse/collections/tools/NoSystemExitRule.java
@@ -1,0 +1,30 @@
+/*
+ * Copyright (c) 2016 Goldman Sachs.
+ * All rights reserved. This program and the accompanying materials
+ * are made available under the terms of the Eclipse Public License v1.0
+ * and Eclipse Distribution License v. 1.0 which accompany this distribution.
+ * The Eclipse Public License is available at http://www.eclipse.org/legal/epl-v10.html
+ * and the Eclipse Distribution License is available at
+ * http://www.eclipse.org/org/documents/edl-v10.php.
+ */
+
+package org.eclipse.collections.tools;
+
+import org.junit.rules.ExternalResource;
+
+public class NoSystemExitRule extends ExternalResource
+{
+    private final ExitMock exitMock = ExitMock.getInstance();
+
+    @Override
+    protected void before()
+    {
+        this.exitMock.setUp();
+    }
+
+    @Override
+    protected void after()
+    {
+        this.exitMock.tearDown();
+    }
+}
